+++ backend/src/neuroagent/app/dependencies.py
@@ -732,6 +732,8 @@ async def get_context_variables(
     entity_frontend_url = settings.tools.frontend_base_url.rstrip("/") + "/app/entity"
     request_id = correlation_id.get()
 
+    storage_frontend_url = settings.tools.frontend_base_url.rstrip("/") + "/app/storage"
+
     return {
         "bluenaas_url": settings.tools.bluenaas.url,
         "bucket_name": settings.storage.bucket_name,
@@ -749,6 +751,7 @@ async def get_context_variables(
         "request_id": request_id,
         "s3_client": s3_client,
         "sanity_url": settings.tools.sanity.url,
+        "storage_frontend_url": storage_frontend_url,
         "shared_state": shared_state,
         "thread_id": thread.thread_id,
         "thumbnail_generation_url": settings.tools.thumbnail_generation.url,