+++ frontend/src/components/chat/chat-messages-inside-thread.tsx
@@ -6,12 +6,12 @@ import {
   getStorageID,
   getValidationStatus,
 } from "@/lib/utils";
-import PlotsInChat from "@/components/chat/plot-in-chat";
 import { ChatMessageAI } from "@/components/chat/chat-message-ai";
 import { ChatMessageHuman } from "@/components/chat/chat-message-human";
 import { ChatMessageTool } from "@/components/chat/chat-message-tool";
 import { ChatMessageLoading } from "./chat-message-loading";
 import { ReasoningCollapsible } from "./reasoning-collapsible";
+import { BackupPlot } from "./backup-plot";
 import { Dispatch, SetStateAction } from "react";
 
 type ChatMessagesInsideThreadProps = {
@@ -59,8 +59,12 @@ export function ChatMessagesInsideThread({
 
   return (
     <>
-      {messages.map((message, idx) =>
-        message.role === "assistant" ? (
+      {messages.map((message, idx) => {
+        const isStreamingLastMsg =
+          (loadingStatus === "streaming" || loadingStatus === "submitted") &&
+          idx === messages.length - 1;
+
+        return message.role === "assistant" ? (
           <div key={message.id}>
             {message.parts.some((part) => part.type === "reasoning") && (
               <ReasoningCollapsible
@@ -69,9 +73,7 @@ export function ChatMessagesInsideThread({
                   ?.filter((part) => part.type === "reasoning")
                   .map((part) => part.reasoning)}
                 messageId={message.id}
-                isReasoning={
-                  !(loadingStatus === "ready") && idx === messages.length - 1
-                }
+                isReasoning={isStreamingLastMsg}
               />
             )}
             {message.parts?.map((part, partId) => {
@@ -85,6 +87,7 @@ export function ChatMessagesInsideThread({
                   message.annotations,
                   part.toolInvocation.toolCallId,
                 );
+                const storageIds = getStorageID(part) || [];
                 return (
                   <div
                     key={`${message.id}-tool-${part.toolInvocation.toolCallId}`}
@@ -102,7 +105,11 @@ export function ChatMessagesInsideThread({
                       simConfigJson={simConfigJson}
                       setSimConfigJson={setSimConfigJson}
                     />
-                    <PlotsInChat storageIds={getStorageID(part) || []} />
+                    <BackupPlot
+                      storageIds={storageIds}
+                      message={message}
+                      isStreamingLastMsg={isStreamingLastMsg}
+                    />
                   </div>
                 );
               }
@@ -120,8 +127,8 @@ export function ChatMessagesInsideThread({
           </div>
         ) : (
           <ChatMessageHuman key={message.id} content={message.content} />
-        ),
-      )}
+        );
+      })}
       {loadingStatus !== "ready" && <ChatMessageLoading />}
     </>
   );