+++ frontend/src/hooks/get-presigned.ts
@@ -1,5 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
 import { useFetcher } from "@/hooks/fetch";
+import { CustomError } from "@/lib/types";
 
 export function useGetPresignedUrl(storageId: string) {
   const fetcher = useFetcher();
@@ -13,5 +14,15 @@ export function useGetPresignedUrl(storageId: string) {
     return response as string;
   };
 
-  return useQuery({ queryKey: [storageId], queryFn: fetchPresignedUrl });
+  return useQuery({
+    queryKey: [storageId],
+    queryFn: fetchPresignedUrl,
+    retry: (failureCount, error) => {
+      if (error instanceof CustomError && error.statusCode === 404) {
+        return false;
+      }
+      return failureCount < 3;
+    },
+    staleTime: 0,
+  });
 }